requirements.txt
fastapi
uvicorn[standard]
pystray
keyboard
pyperclip
qrcode[pil]
pillow